renaming some files to be consistent with other plugins
